refactor: utils to calendar folder
